requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudscraper==1.2.71
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
gunicorn==23.0.0
h11==0.16.0
idna==3.10
mercadopago==2.3.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
whitenoise==6.9.0
